requirements.txt
anyio==3.7.0
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
cryptography==41.0.1
exceptiongroup==1.1.1
fastapi==0.97.0
h11==0.14.0
idna==3.4
mysqlclient==2.1.1
passlib==1.7.4
pycparser==2.21
pydantic==1.10.9
pymodbus==3.3.2
pyModbusTCP==0.2.0
PyMySQL==1.0.3
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.17
sqlalchemy2-stubs==0.0.2a34
starlette==0.27.0
typing_extensions==4.6.3
uvicorn==0.22.0
